Rename gp_mix module
